make syncAllWithRemote() a class method of Widget instead of AppDelegate
